requirements.txt
disnake
python-dotenv
aiohttp
